% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlsampler.R
\name{dl.beta}
\alias{dl.beta}
\title{Code for paper is given at https://github.com/debdeeptamu/Dirichlet_Laplace/blob/master/DL.m}
\usage{
dl.beta(XtX, Xty, sig.sq, tau.sq, psi, phi)
}
\description{
Code for paper is given at https://github.com/debdeeptamu/Dirichlet_Laplace/blob/master/DL.m
}
