% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlsampler.R
\name{sym.sq.root}
\alias{sym.sq.root}
\title{Code for paper is given at https://github.com/debdeeptamu/Dirichlet_Laplace/blob/master/DL.m}
\usage{
sym.sq.root(A)
}
\description{
Code for paper is given at https://github.com/debdeeptamu/Dirichlet_Laplace/blob/master/DL.m
}
